cli/cluster: wire --quiet on logs/yaml verbs, lift sleep + events helpers into clusteropts
Two real UX bugs and four high-confidence dedup wins:

A1/B1: pod logs, container logs, and the four `* yaml` verbs all called
       o.WriteStdout / o.PrintJSON which are documented to honor --quiet,
       but none of them registered the flag — so --quiet was unreachable.
       Added a small AddQuietFlag helper (sibling to AddOutputFlags) and
       wired it on all six verbs.

A2:    `container env -o json` against an empty-container pod yielded zero
       stdout because the stderr-and-return path ran before the JSON branch.
       Reordered so JSON mode always emits {containers: []} (parseable by
       scripts), table mode prints the stderr message, and --container=<x>
       not-found stays a hard error in either output mode.

B2:    Three identical sleepCtx copies in pod/logs.go, workload/rollout.go,
       and application/status.go collapsed into clusteropts.SleepContext.
       pod/get.go's piggy-back on the pod-package copy redirects to the
       shared symbol too.

B3:    Event type, sort key, and table renderer were copy-pasted across
       pod/events.go, job/events.go, and (for type + sort key)
       application/status.go. Pulled them into clusteropts/events.go.
       job/events.go no longer needs `import pod`. application/status.go
       keeps its own per-section table renderer (different column set,
       indented, embedded in a larger layout) but now consumes the shared
       Event type + EventSortTime sort key.

B4:    The `/api/v1/namespaces/<ns>/events[?limit=N]` URL was built three
       different ways at three call sites; replaced with one
       clusteropts.EventsListPath helper.

Net diff: 14 files changed, 64 insertions, 269 deletions (~-205 LOC).

Build, vet, race-tests on clusteropts, and lints all green. Help text
confirms --quiet is now visible on all six wired verbs.

Made-with: Cursor
