+++ cli/cmd/ctl/cluster/cronjob/yaml.go
@@ -43,6 +43,7 @@ point). For JSON, use ` + "`cluster cronjob get -o json`" + `.
 		},
 	}
 	cmd.Flags().StringVarP(&namespace, "namespace", "n", "", "namespace (required when the positional argument is a bare name)")
+	o.AddQuietFlag(cmd)
 	return cmd
 }
