+++ cli/cmd/ctl/cluster/pod/yaml.go
@@ -54,6 +54,7 @@ point of the verb). For JSON, use ` + "`cluster pod get -o json`" + `.
 		},
 	}
 	cmd.Flags().StringVarP(&namespace, "namespace", "n", "", "namespace (required when the positional argument is a bare name)")
+	o.AddQuietFlag(cmd)
 	return cmd
 }
