+++ cli/skills/olares-cluster/SKILL.md
@@ -131,7 +131,7 @@ K8s Jobs (`apis/batch/v1`).
 | `cluster job get <ns/name \| name> [-n NS]` | `GET /apis/batch/v1/namespaces/<ns>/jobs/<name>` | Vertical summary including Active / Succeeded / Failed counts, Conditions list, and a `Controlled By: CronJob/<name>` line when the job has a CronJob owner reference. Exports `job.Get` for `job pods`/`rerun` to share. |
 | `cluster job yaml <ns/name \| name> [-n NS]` | same endpoint | JSON-to-YAML round-trip. |
 | `cluster job pods <ns/name \| name> [-n NS] [-l ADDITIONAL] [--field-selector SEL] [--limit N]` | (1) `job.Get` for `metadata.uid`; (2) `pod.RunList` with `labelSelector=controller-uid=<uid>[,<extra>]` | Two-step "find pods this job spawned". Mirrors the SPA's lazy-load tree. `--label` is ANDed onto the controller-uid clause server-side. Server-decides scoping. |
-| `cluster job events <ns/name \| name> [-n NS] [--limit N]` | `GET /api/v1/namespaces/<ns>/events` | Same shape as `cluster pod events` filtered to `involvedObject.kind=Job, name=<job>`. Reuses `pod.Event` typed view. Sorted oldest-first by `lastTimestamp`. |
+| `cluster job events <ns/name \| name> [-n NS] [--limit N]` | `GET /api/v1/namespaces/<ns>/events` | Same shape as `cluster pod events` filtered to `involvedObject.kind=Job, name=<job>`. Shares the `clusteropts.Event` typed view, `RenderEventsTable`, and `EventsListPath` URL builder with `cluster pod events` and `cluster application status`. Sorted oldest-first by `lastTimestamp`. |
 | `cluster job rerun <ns/name \| name> [-n NS] [--yes]` | (1) `job.Get` for `metadata.resourceVersion`; (2) `POST /kapis/operations.kubesphere.io/v1alpha2/namespaces/<ns>/jobs/<name>?action=rerun&resourceVersion=<rv>` (no body) | KubeSphere operations action — server spawns a new pod attempt and updates `Status.Active` accordingly. `ConfirmDestructive`-wrapped (rerun is mutating). |
 
 ### CronJobs (`cluster cronjob ...`, aliases `cronjobs` / `cj`)
@@ -241,13 +241,13 @@ Every mutating verb in this tree follows the same template. When adding a new on
 | [`cli/cmd/ctl/cluster/internal/clusteropts/argparse.go`](cli/cmd/ctl/cluster/internal/clusteropts/argparse.go) | `SplitNsName(nsFlag, arg)` — single source for the `<ns>/<name>` or `-n + name` positional grammar shared by every per-noun get/yaml/pods/events verb. |
 | [`cli/cmd/ctl/cluster/internal/clusteropts/format.go`](cli/cmd/ctl/cluster/internal/clusteropts/format.go) | `DashIfEmpty(s)` and `Age(ts, now)` — single-source row-rendering helpers shared by every list/get table renderer. |
 | [`cli/cmd/ctl/cluster/internal/clusteropts/yaml.go`](cli/cmd/ctl/cluster/internal/clusteropts/yaml.go) | `JSONToYAML(body)` — single source for the K8s native JSON-to-YAML conversion used by every `cluster <noun> yaml` verb. |
-| [`cli/cmd/ctl/cluster/pod/`](cli/cmd/ctl/cluster/pod) | `cluster pod` verbs (`list`, `get` with `--watch`, `yaml`, `events`, `logs`, `delete`, `restart`); exports `pod.Get` + `pod.RunList` + `pod.RunLogs` + `pod.LogsOptions` + `pod.RunDelete` + `pod.Event` for sibling packages. |
+| [`cli/cmd/ctl/cluster/pod/`](cli/cmd/ctl/cluster/pod) | `cluster pod` verbs (`list`, `get` with `--watch`, `yaml`, `events`, `logs`, `delete`, `restart`); exports `pod.Get` + `pod.RunList` + `pod.RunLogs` + `pod.LogsOptions` + `pod.RunDelete` for sibling packages. The shared `Event` typed view + sort/render helpers live in `clusteropts/events.go`. |
 | [`cli/cmd/ctl/cluster/container/`](cli/cmd/ctl/cluster/container) | `cluster container` verbs (`list`, `env`, `logs`). `list`/`env` are pure projections over `pod.Get`; `logs` is a thin alias over `pod.RunLogs` that adds the 3-segment `<ns>/<pod>/<container>` positional grammar. |
 | [`cli/cmd/ctl/cluster/workload/`](cli/cmd/ctl/cluster/workload) | `cluster workload` verbs (`list`, `get`, `yaml`, `rollout-status`, `scale`, `restart`, `stop`, `start`, `delete`); covers Deployment + StatefulSet + DaemonSet via per-call `--kind`. Exports `workload.RunList`, `workload.NormalizeKind`, `workload.SingularKind`, `workload.RunScale`. `rollout-status` and `scale --watch` share a single convergence loop (no duplicated polling code). `restart` is SPA-aligned (GET selector → list pods → bounded-parallel DELETE). |
 | [`cli/cmd/ctl/cluster/application/`](cli/cmd/ctl/cluster/application) | `cluster application` verbs (`list`, `get`, `workloads`, `pods`, `status`); workloads/pods delegate to the sibling packages. `status` is CLI-original — parallel `errgroup` fan-out across the underlying GETs with three-section table / one-object-per-tick JSONL. |
 | [`cli/cmd/ctl/cluster/namespace/`](cli/cmd/ctl/cluster/namespace) | `cluster namespace` verbs (`list`, `get`) — K8s framing of the same resource as `application`. |
 | [`cli/cmd/ctl/cluster/node/`](cli/cmd/ctl/cluster/node) | `cluster node` verbs (`list`, `get`). Per-user K8s view; not the host-side `olares-cli node` tree. |
-| [`cli/cmd/ctl/cluster/job/`](cli/cmd/ctl/cluster/job) | `cluster job` verbs (`list`, `get`, `yaml`, `pods`, `events`, `rerun`). Exports `job.Get` for siblings. `pods` delegates to `pod.RunList` with `controller-uid=<uid>`; `events` reuses `pod.Event`; `rerun` POSTs to the KubeSphere operations API. |
+| [`cli/cmd/ctl/cluster/job/`](cli/cmd/ctl/cluster/job) | `cluster job` verbs (`list`, `get`, `yaml`, `pods`, `events`, `rerun`). Exports `job.Get` for siblings. `pods` delegates to `pod.RunList` with `controller-uid=<uid>`; `events` reuses `clusteropts.Event` + the shared render/sort/URL helpers; `rerun` POSTs to the KubeSphere operations API. |
 | [`cli/cmd/ctl/cluster/cronjob/`](cli/cmd/ctl/cluster/cronjob) | `cluster cronjob` verbs (`list`, `get`, `yaml`, `jobs`, `suspend`, `resume`). Exports `cronjob.Get` for siblings. `jobs` delegates to a K8s native `apis/batch/v1/.../jobs?labelSelector=…` GET (rebuilt from `spec.jobTemplate.metadata.labels`); `suspend`/`resume` PATCH `application/merge-patch+json` and share `runToggle`. |
 | [`cli/cmd/ctl/cluster/middleware/`](cli/cmd/ctl/cluster/middleware) | `cluster middleware` verbs (`list`, plus the `password` sub-noun). Custom envelope; passwords redacted by default. |
 | [`cli/cmd/ctl/cluster/middleware/password/`](cli/cmd/ctl/cluster/middleware/password) | `cluster middleware password set` — `POST /middleware/v1/<type>/password` with no-echo prompt + `ConfirmDestructive`. |
